+++ python/lsst/meas/transiNet/__init__.py
@@ -22,4 +22,4 @@
 from .version import *  # Generated by sconsUtils
 
 from .rbTransiNetInterface import *
-from .transiNetTask import *
+from .rbTransiNetTask import *
